auth-node: move identity related modules to identity dir
