+++ plugins/auth-node/src/identity/getBearerTokenFromAuthorizationHeader.test.ts
