+++ plugins/auth-node/src/identity/getBearerTokenFromAuthorizationHeader.ts
